Remove hugo urls
There are a number of pages which had calls back to localhost:1313,
which causes issues when trying to read the docs online.

I've gone through and done a mass grep for localhost:1313 to fix these
in the content/ folder.

Also fixed the CQ link on the downsampling and retention page.